Move shared middleware to middleware folder
Consistency is helping me navigate code
