+++ src/shared/middleware/__tests__/make-app-engine-request-id-middleware.test.js
@@ -1,10 +1,10 @@
 //@flow
-import * as GetRequestLogger from "../get-request-logger.js";
+import * as GetRequestLogger from "../../get-request-logger.js";
 import {makeAppEngineRequestIDMiddleware} from "../make-app-engine-request-id-middleware.js";
-import * as GetAppEngineRequestID from "../get-app-engine-request-id.js";
+import * as GetAppEngineRequestID from "../../get-app-engine-request-id.js";
 
-jest.mock("../../ka-shared/index.js");
-jest.mock("../get-app-engine-request-id.js");
+jest.mock("../../../ka-shared/index.js");
+jest.mock("../../get-app-engine-request-id.js");
 
 describe("#makeErrorMiddleware", () => {
     it("should return middleware function", () => {
